refactor(fabric-async): :sparkles: export types formerly defined in base module
