+++ packages/fabric-async/src/events/event-gamify.ts
@@ -1,6 +1,5 @@
 import {
   BaseEvent,
-  ContextSpace,
   EventName,
   RecordOperation,
 } from "../base"
@@ -8,6 +7,10 @@ import {
 import {
   Leaderboard,
 } from "../activity"
+
+import {
+  ContextSpace,
+} from "../context"
 
 export interface EventLeaderboard extends BaseEvent<
   EventName<"leaderboard", RecordOperation>,
