+++ packages/fabric-async/src/events/event-purchase.ts
@@ -4,10 +4,14 @@ import {
 
 import {
   BaseEvent,
-  ContextSession,
   EventName,
-  PurchaseOperation,
 } from "../base"
+
+import {
+  ContextSession,
+} from "../context"
+
+export type PurchaseOperation = "payment_completed" | "payment_refunded"
 
 export interface EventPurchase extends BaseEvent<
   EventName<"purchase", PurchaseOperation>,
