+++ packages/fabric-async/src/events/event-reward.ts
@@ -1,13 +1,17 @@
 import {
   BaseEvent,
-  ContextUserActivity,
   EventName,
-  RewardOperation,
 } from "../base"
+
+import {
+  ContextUserActivity,
+} from "../context"
 
 import {
   Reward,
 } from "../activity"
+
+export type RewardOperation = "provisioned" | "fulfilled"
 
 export interface EventReward extends BaseEvent<
   EventName<"reward", RewardOperation>,
