+++ packages/fabric-async/src/events/event-session.ts
@@ -1,9 +1,13 @@
 import {
   EventName,
-  SessionOperation,
   BaseEvent,
-  ContextSession,
 } from "../base"
+
+import {
+  ContextSession,
+} from "../context"
+
+export type SessionOperation = "signup" | "signin" | "entered_space"
 
 export interface EventSession extends BaseEvent<
   EventName<"session", SessionOperation>,
