+++ packages/fabric-async/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./activity"
+export * from "./context"
 export * from "./events"
+export * from "./base"
